requirements.txt
kafka-python==2.0.1
numpy==1.19.1
pandas==1.1.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
psycopg2==2.8.5
